tests/wpt/web-platform-tests/css/css-flexbox/flex-one-sets-flex-basis-to-zero-px.html

<!DOCTYPE html>
<html>
<title>CSS Flexbox: flex-basis with zero pixel</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#flex-basis-property">
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />
<link href="support/flexbox.css" rel="stylesheet">
<meta name="assert" content="This test ensures that setting 'flex-basis' property to
zero pixel or percentage in combination with one value of both 'flex-grow' and 'flex-shrink'
and 'writing-mode' value works properly.">
<style>
.flexbox > div {
    font: 14px/1 Ahem;
}

.flex-one-one-zero-percent {
    flex: 1 1 0%;
    min-width: 0;
    min-height: 0;
}

.flex-one-one-zero-px {
    flex: 1 1 0px;
    min-width: 0;
    min-height: 0;
}

.vertical {
    writing-mode: vertical-rl;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="document.fonts.ready.then(() => { checkLayout('.flexbox'); })">
<div id=log></div>
<div class="flexbox column">
  <div class="flex-one" data-expected-height="14">Flex item with flex: 1</div>
</div>

<div class="flexbox column">
  <div class="flex-one-one-zero-percent" data-expected-height="14">Flex item with flex: 1 1 0%</div>
</div>

<div class="flexbox column">
  <div class="flex-one-one-zero-px" data-expected-height="0">Flex item with flex: 1 1 0px</div>
</div>

<div class="flexbox column vertical">
  <div class="flex-one" data-expected-width="14">Flex item with flex: 1</div>
</div>

<div class="flexbox column vertical">
  <div class="flex-one-one-zero-percent" data-expected-width="14">Flex item with flex: 1 1 0%</div>
</div>

<div class="flexbox column vertical">
  <div class="flex-one-one-zero-px" data-expected-width="0">Flex item with flex: 1 1 0px</div>
</div>
</body>
</html>
